Reorganize dir structure of thirdparty.
